[ui] Check the stylesheets structurally instead of snapshotting them
The golden table pinned all 68 rendered constants. That was worth having while
the 116 hex literals were being converted to token references -- it is how the
conversion was shown to be inert rather than asserted to be -- but as a standing
test it guards the wrong thing. Accidentally editing a colour is a risk that
predates the f-strings and that nothing else was testing; the hazard the
f-strings actually introduced is structural.

    QProgressBar {{ color: {TEXT_COLOR}; }}      ->  color: #d6d6d6
    QProgressBar {{ color: {{TEXT_COLOR}}; }}    ->  color: {TEXT_COLOR}

Both are valid Python, the second raises nothing, and Qt drops a rule it cannot
parse without saying so. Checking for the leftover placeholder catches that
directly, with nothing to regenerate after a deliberate restyle -- and the table
was also what made this test fail on CI for a reason that was not a style
regression at all.

So: drop _GOLDEN and the per-constant comparison, keep the import smoke test,
the placeholder check and the absolute-path check. Renamed, because there is no
golden data in it any more.

Verified all three failure modes still fail: an over-doubled brace trips the
placeholder check, an undoubled one trips the import, and a leaked absolute path
trips the path check.
